requirements.txt
faiss-cpu
openai
aioboto3
numpy